test: split change-graph fold tests into their own topic file (ADR 0028)
Move the ADR 0012 decision 1 "— uses: ..." folding tests (root
exemption, per-parent repetition, non-fold-on-cycle-edge exceptions,
and the HCL fold test) out of change_graph_tree.rs (1080 lines, over
the 1000-line warn threshold) into a new sibling topic file
change_graph_fold.rs, keeping true-pre-order-DFS/(see above)/cycle-
warning tests in the original file.

Also fixes a review finding in the moved HCL fold test: its two
parent symbols are real HCL resource blocks, which extraction always
reports as SymbolKind::Block, never SymbolKind::Function — update
both to SymbolKind::Block and pin the resulting `block`-prefixed
Markdown output.
